refactored app add enum for routs and increase version of translate module
