+++ .prettierrc.json
@@ -27,11 +27,11 @@
 
     "^@shared/(.*)$",
 
+    "^@utils/(.*)$",
+
     "^@assets/(.*)$",
 
     "^@env/(.*)$",
-
-    "^utils/(.*)$",
 
     "^@app/(.*)$",
