+++ package.json
@@ -5,7 +5,6 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "deploy": "ng build --configuration production",
     "watch": "ng build --watch --configuration development",
     "lint": "ng lint",
     "prettier": "npx prettier . --write",
@@ -37,7 +36,7 @@
     "@ngrx/router-store": "^16.3.0",
     "@ngrx/store": "^16.2.0",
     "@ngrx/store-devtools": "^16.3.0",
-    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/core": "^15.0.0",
     "@ngx-translate/http-loader": "^7.0.0",
     "@types/firebase": "^3.2.1",
     "dom-to-image": "^2.6.0",
