+++ src/app/core/components/dark-mode-selector/dark-mode-selector.component.scss
@@ -11,11 +11,11 @@
     visibility: hidden;
   }
 
-  &__input:checked+label {
+  &__input:checked + label {
     background: #242424;
   }
 
-  &__input:checked+label:after {
+  &__input:checked + label:after {
     left: 48.5px;
     transform: translateX(-100%);
     background: url('../../../../assets/img/dark-mode-button/moon.svg');
