+++ src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-import { firebaseConfig } from 'utils/firebase/firebase.config';
+import { firebaseConfig } from '@utils/firebase/firebase.config';
 
 export const environment = {
   production: true,
