+++ src/utils/enum/router-path.enum.ts
@@ -1,0 +1,10 @@
+export enum ERouterPath {
+  AUTH = 'auth',
+  LAYOUT = 'layout',
+  ABOUT_ME = 'about-me',
+  EDUCATION = 'education',
+  PROJECTS = 'projects',
+  TECHNOLOGIES = 'technologies',
+  WORK_EXPERIENCE = 'work-experience',
+  CV = 'cv'
+}
